Switch to C++ exception handling
This removes setjmp/longjmp exception handling and replaces it with C++
exception handling.  The most important outcome of this is that the
automatic reference counting stuff (e.g., for auto GC mode) doesn't
leak references anymore.

Another benefit of this change is that exception handling is now
(basically) threadsafe.  The one part which is somewhat questionable is
the handling of uncaught exceptions, which still uses the global
Rogue_error_object object.  When the uncaught exception handler runs,
this *hopefully* contains "the most recent exception", but this can't
really be relied upon.  Also, this functionality was moved into its own
function, and it's now set up in Rogue_configure() instead of
Rogue_launch().

Both uncaught exceptions as well as the segfault handler now result in
the process exiting with exit code 1.