Exception: Set Exception message by default
This lets you skip setting a specific message, and the message gets
set to the name of the Exception subclass, which is reasonable in
some cases.

This also used to avoid a case where you could get an Exception
where the message was not set, since the compiler let you create
an instance as if it had a default constructor even when it didn't,
but I think that may be fixed now anyway.