CPPWriter: normalize select expressions
C++ requires that the second and third arguments to the conditional
operator not be convertible to each other (if they can be, it's not
clear which type the whole expression should be).  In some cases
with reference types, this wasn't happening, because you might get
one argument which is a raw pointer and one which is a smart pointer
(for GC), and there are conversions both ways.

This commit fixes this by wrapping all the second and third
arguments in ROGUE_ARG, converting non-smart pointers to smart ones.
(It's possible that doing the opposite would also work...)