NativeCPP: Fix exceptions with --gc=auto
When throwing, we may be given a GC-wrapped exception pointer
which the catch statements will never catch.  We now cast it
to a raw pointer, which should fix things.  I'm not sure, but
losing the GC wrap here may be significant... BUT, the catches
should all re-wrap anyway (via _internal_exception_reference),
so it shouldn't be a problem.