Fix auto GC with modified parameter values
Values always have a nonzero reference count when they're passed as
parameters, so there's no need to actually pass references as
RoguePtrs -- they can never be freed until the callee returns anyway,
and not doing so is cheaper (at the expense of possibly retaining
some memory which could theoretically be freed earlier).

*Except* that there's a previously-unaccounted-for gotcha here.  Since
parameters are raw pointers, if one assigns a *new* value to a
parameter (treating it as if it were a variable), the new value isn't
GC-managed since it doesn't end up in a RoguePtr.  This commit fixes
this case.