Add "IntPtr" definition
The compiler now defines a macro "IntPtr" which evaluates to an Int that's
large enough to hold a pointer (in fact, the underlying intptr_t type).
This is basically the same as what the Cython plugin used to do (though
it used to be called PointerInt), so that code has been removed.