StringBuilder: Make work buffer threadLocal
This commit uses conditional compilation to make the work buffer
threadLocal when compiling with multithreading or not threadLocal
otherwise.  This shouldn't really be necessary -- if not compiling
for threading, [threadLocal] should have no real effect.  The
reason it's done here is for the sake of bootstrapping, since
old versions of roguec would choke on the property attribute.
Someday we can get rid of the conditional compilation.