Cython: Expose methods on Global globally
Previously, to access members of the Rogue Global object from
Python, you had to do so explicitly (mod.Global.foo()).  But
now that routines are methods on Global, this meant that you
had to do so for routines, which seemed unfortunate.  This
commit exposes methods on Global as the module level in
Python.  In particular, this means routines act like they
used to (mod.foo() works again for a routine named foo).