Improve return values for MethodInfo invocation
Previously, when invoking a method via MethodInfo, you were required to
use the .returns<<...>> template method to specify the return type.  Abe
had the idea of  using the '->' operator, which is consistent with, e.g.,
native(...)->type.  So we now do this.  Internally, it works by just
translating to the .returns method.

Additionally, if you try to use the return value of a .call invocation,
the compiler gives a special error message, which is more meaningful
than what you'd get without the special handling.

Finally, we improve the runtime return type checking in several ways:
* Checks are always done at runtime.  If it's a performance problem
  we could consider making it only in DEBUG builds.
* We now check that the non-return-value versions are invoking
  methods that actually have no return value.
* On ones that return a value, we use is_subclass to check the
  return type, rather than an exact match.
* We generate fairly helpful error messages on failure.