Fixed bug in auto-generated introspection call framework.
For introspection, unrelated methods are grouped as long as they have
compatible signatures.  The return type of the first method in a group
was arbitrarily being saved as a placeholder return type, but if that
type was culled out by the end of the program the placeholder return
type wasn't being updated.