NativeCPP: Fix RogueArrays with GCC
RogueArray has a union full of flexible array members, which I think
is actually against the standard, isn't it?  clang seems to let it
slide, but GCC doesn't.  This kept it from compiling on GCC.  On the
other hand, GCC allows zero-length arrays, which I think is also
against the standard, but works for us here.  So basically, we use
the former odd behavior on clang and the latter odd behavior on GCC.

FWIW, I think the real solution here might be to have RogueArray be
a template specialized on the array type rather than a union of all
possible types.