mtgc: Add the auto-mt threadsafe GC
This adds a threadsafe version of the --gc=auto collector.  It
is sort of a variant of a "stop the world" collector.  Maybe more
like a "wait for the world to stop" collector.  GC is run on its
own dedicate thread.  The compiler generates code which, from
time to time, has each thread check to see whether GC should be
performed.  If so, the thread pauses and signals the GC thread that
it's okay to GC  as far as that thread is concerned.  When all
threads are paused, the GC runs and then unpauses the threads.
When the threads offer to pause, they are at a "safe spot" to GC --
they don't have any un-retained references or aren't in the middle
of doing an allocation, etc.