mt: Synchronize singleton initialization
Singletons are initialized on demand, which could previously lead to a
race in multithreaded programs where different threads initialize the
same singleton at once.  In MT modes, the singleton reference is now an
atomic pointer, and initialization happens under a recursive lock
(recursive so we don't deadlock in case one singleton causes another
singleton to be initialized).