CPPWriter: Fix a comparison expression issue
There was a pretty substantial issue with some expressions, e.g.:
  local v1 = 2
  local v2 = 2
  println (v1 != v2) != false # false
  println false != (v1 != v2) # true

The previous commit was a start at fixing the issue.  This commit
finishes it.

The issue stems from trying to avoid an annoying clang warning about
conditionals with extra parentheses.  This commit wraps conditionals
in a no-op macro, which silences the warning.  This is kind of ugly,
but we decided it was preferable to a) doing a bunch of extra work
to sometimes generate the parentheses and sometimes not or b) needing
needing to keep track of another compiler flag that will vary between
compilers and compiler versions.