[Introspection] Fixed severe introspection call bug.
Each generated introspection call handler can have many methods mapped
to it. Any parameter types and return types that are instanceOf 'Object'
(except for 'Value') were mapped to a handler with a corresponding
'Object' parameter or return type.

The problem was that String was also mapped to Object parameter types
and return types but was handled differently by the introspection call
handler code generator, and so there were situations were an object
would end up getting converted to a StringValue and back again rather
than being wrapped in an ObjectValue.

The fix is that String types are treated as their own category and are
no longer mapped to Object types when mapping to introspection call
handlers.