stdlib: Process reads from process before it exits
Previously, Process read from the child only after the child had exited.
This only works if the process never blocks waiting for you to read
from it (i.e., if the amount of output data is small).  We now read
until there's nothing more to read and *then* wait for it to exit.