NativeCPP: Fix introspection type info reading
fcc110c522c6c254 seems to have contained a regression which caused it to
no longer read the method_count type info when compiled without
introspection.  However, this type_info is written regardless of whether
introspection is enabled, so it needs to be read regardless.  This
would lead to any following type info being read wrong, which hasn't
been a problem *yet* since this is currently the final info.  But
it will be a problem next time any new type info is added.