roguec: Fix issue where template method inferred more than once
When a method template is specialized via inference, we build a new
Method, then check to see if the type already has a matching method,
and only inject the new method into the type if there isn't.

Previously, we would then always return the new Method.  This led to
a compiler crash, presumedly because that Method never actually ended
up being part of any type.  We now return the existing method if
there was one.

It's possible there are still some problematic corner cases (e.g.,
where there's a global method with the same signature but not an
instance method or something like that).