python: Make thisptr a _RogueObjectPtr for new instances
There were at least some cases where the thisptr of a Python wrapper
type would end up being a Python int instead of a _RogueObjectPtr
(AKA ctypes void pointer).  This probably works fine in a lot of cases,
but it's slightly confusing and could be annoying if you, e.g.,
expected to be able to access its .value property.