python: Make comparison operators work
The existing code was from the Cython implementation and didn't actually
work with ctypes.  We now pretty much just translate the names and
they work right, though the the overloading mechanism needed slight
tweaking to work with double-underscored names which are slightly
magical in Python.

We also now use the total_ordering decorator on all classes so that
we hopefully get all comparisons even if only a subset are actually
implemented.  With all the usual total_ordering caveats.