[List] Renamed mapped(), reduced() -> map(), reduce()
The past-tense naming convention is only used to denote a non-mutating method that produces a List or object of the same type as the context class. shuffle() mutates the list while shuffled() returns a cloned and shuffled list.

If a method produces an object of a different class then the infinitive form should be used, e.g. map() instead of mapped().