[Process] Windows Process quoting tweak and annotation
Using 'cmd.exe /c' to run programs is a bit of a hack, more so since the command line needs an extra pair of double-quotes around everything.

The initial implementation added double-quotes to args after the args list was already built, even if the args list does not contain 'cmd.exe', '/c'.

Tweaked to only add double-quotes to the command line when 'cmd.exe', '/c' is being used.

Also added the following note:

I forget if there's a specific reason I chose 'cmd.exe /c' - perhaps just to do a path search.  If there are any troubles with this approach, might be worth doing path searching manually (calling the Windows API function to do a path search) and requiring a caller that wants to run a cmd.exe built-in to specify cmd.exe themselves. --Abe