[String] Added .contains_pattern(), .locate_pattern(), .matches_pattern(), and $($) escape.
Uses the same patterns as .extract_string() and .extract_strings().  Here's the run-down:

The format pattern is matched literally apart from the following special sequences and behavior:
  *    - Matches any number of characters.
  ?    - Matches any single character.
  $    - Like '*' but discards leading whitespace; is used as a marker in extract_string(s).
  $(I) - Matches an integer character sequence.
  $(R) - Matches a real number character sequence.
  $($) - Matches a single literal '$'.
  ' '  - A space matches 1 or more spaces and tabs.

Finally, extra whitespace in this string that isn't matched in the pattern string is ignored as long as the whitespace doesn't have alphanumeric characters on both sides.