python: Set restype and argtypes on correct functions
We generated wrapped versions of functions which handle exceptions.
These are the ones which actually get called from Python.  The
corresponding ctypes function objects need to have their return
types and arguments set, but we were previously setting them on the
*original* versions of the functions -- not the wrapped ones --
so all the calls to the wrappers were with the wrong types.