python: Attempt to retain objects held by Python
Most objects held by Python should already be properly retained, but
there's at least one exception, which has to do with Strings.  Python
generally doesn't hold Rogue Strings -- it holds Python strings.  When
it actually needs to pass a String to Rogue, it creates a Rogue String
temporarily.  Previously, this wasn't actually guaranteed to be
retained, so it's *possible* it'd be freed before Rogue did something
with it.  Probably not, but maybe.  We now explicitly hold references
to these on the Python side.  We do something similar with delegates
because I think they may also possibily have this problem.