python: Add compatibility with PyPy
When compiling a Python module, you can pass PYROGUE_PYPY_COMPATIBLE,
and the resulting module will be usable by Python 2, Python 3, and PyPy
(only tested with PyPy 2 so far, but should work with PyPy 3 as well).
In other words, such a module should work for basically any Python.
It *might* be slightly slower on CPython, though, which is why we still
support the older, non-compatible method as well.

There are basically two things involved:
* To retain Python objects when they're being held by Rogue, we  add
  them to a dictionary of retained objects rather than using the
  Py_IncRef/Py_DecRef type functions directly.
* When Rogue code wants to raise an exception in Python, it no longer
  uses the Python C API, and instead uses ctypes' errcheck feature.